requirements.txt
Flask==3.0.3
bcrypt==4.2.0
Flask-Limiter==3.8.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn